Add Client CRUD

+++ server/src/controllers/CategoriesController.ts
@@ -3,7 +3,7 @@
   import Slugify from 'slugify';
   
   import db from '../database/connection';  
-  import validation from '../validations/inputDataValidation';
+  import Validator from '../validations/InputDataValidation';
 
   interface category{
     categoryName: string
@@ -15,6 +15,8 @@ export default {
   async index(req: Request, res:Response){
     try {
       const categories:Array<category> = await db.select().from('categories');
+      if(categories.length === 0) return res.sendStatus(204);
+
       res.status(200).json(categories);
     } catch (error) {
       console.log(error);
@@ -29,8 +31,10 @@ export default {
         categoryName: categoryName,
         categoryUrl: Slugify(categoryName)
       }
-      await validation.categoriesValidation.validate(data);
+
+      await Validator.categoriesValidation.validate(data);
       await db.insert(data).into('categories');
+      
       res.sendStatus(201);
     } catch (error) {
       error.name === 'ValidationError' ? res.sendStatus(400) : res.sendStatus(500);
@@ -42,8 +46,15 @@ export default {
     try {
       const { categoryUrl } = req.params;
       const data:category = req.body;
-      await validation.categoriesValidation.validate(data);
-      await db.update(data).table('categories').where({categoryUrl: categoryUrl});
+
+      const category = await db.select()
+      .from('categories').where({categoryUrl: categoryUrl});
+      if(category.length === 0) return res.sendStatus(404);
+
+      await Validator.categoriesValidation.validate(data);      
+      await db.update(data).table('categories')
+      .where({categoryUrl: categoryUrl});
+
       res.sendStatus(200);
     } catch (error) {
       error.name === 'ValidationError' ? res.sendStatus(400) : res.sendStatus(500);
@@ -54,8 +65,14 @@ export default {
   async deleteCategory(req: Request, res:Response){
     try {
       const { categoryUrl } = req.params;
+
+      const category = await db.select()
+      .from('categories').where({categoryUrl: categoryUrl});
+      if(category.length === 0) return res.sendStatus(404);
+
       await db.delete().from('categories').where({categoryUrl : categoryUrl});
       res.sendStatus(200);
+
     } catch (error) {
       res.sendStatus(400);
       console.log(error);
