+++ server/src/controllers/ProductsController.ts
@@ -1,27 +1,15 @@
 //Imports
   import {Request, Response} from 'express';
-  import axios from 'axios';
   
   import db from '../database/connection';  
-  import validation from '../validations/inputDataValidation';
-
-  async function getAddress(cep: number){
-    try {
-      const address = await axios.get(`https://viacep.com.br/ws/${cep}/json/`);
-      return address.data;
-    } catch (error) {
-      console.log(error);
-      return error;
-    }
-  }
-  getAddress(13218830);
-
+  import Validator from '../validations/InputDataValidation';
+  import Formatter from '../utils/Formatter';
+  
   interface product{
     productName: string,
     quantity: number,
     price: number,
-    purchaseDate: string,
-    category: string,
+    categoryName: string,
     description: string
   };
 
@@ -30,6 +18,8 @@ export default {
   async index(req: Request, res:Response){
     try {
       const products:Array<product> = await db.select().from('products');
+      if(products.length === 0) return res.sendStatus(204);
+
       res.status(200).json(products);
     } catch (error) {
       console.log(error);
@@ -40,7 +30,10 @@ export default {
   async show(req: Request, res:Response){
     try {
       const {id} = req.params;
+
       const product = await db.select().from('products').where({id: id});
+      if(product.length === 0) return res.sendStatus(404);
+
       res.status(200).json(product);
     } catch (error) {
       console.log(error);
@@ -51,8 +44,12 @@ export default {
   async getProductByCategory(req: Request, res:Response){
     try {
       const {categoryUrl} = req.params;
-      const categoryName = categoryUrl.replace('-', ' ');
-      const products:Array<product> = await db.select().from('products').where({categoryName: categoryName});
+
+      const categoryName = Formatter(categoryUrl);
+      const products:Array<product> = await db.select()
+      .from('products').where({categoryName: categoryName});
+      if(products.length === 0) return res.sendStatus(404);
+
       res.status(200).json(products);
     } catch (error) {
       console.log(error);
@@ -63,7 +60,7 @@ export default {
   async createProduct(req: Request, res:Response){
     try {
       const data:product = req.body;
-      await validation.productValidation.validate(data);
+      await Validator.productValidation.validate(data);
       await db.insert(data).into('products');
       res.sendStatus(201);
     } catch (error) {
@@ -75,10 +72,15 @@ export default {
 
   async updateProduct(req: Request, res:Response){
     try {
-       const { id } = req.params;
+      const { id } = req.params;
       const data:product = req.body;
-      await validation.productValidation.validate(data);
+
+      const product = await db.select().from('products').where({id: id});
+      if(product.length === 0) return res.sendStatus(404);
+      
+      await Validator.productValidation.validate(data);
       await db.update(data).table('products').where({id: id});
+
       res.sendStatus(200);
     } catch (error) {
       error.name === 'ValidationError' ? res.sendStatus(400) : res.sendStatus(500);
@@ -89,8 +91,13 @@ export default {
   async deleteProduct(req: Request, res:Response){
     try {
       const { id } = req.params;
+
+      const product = await db.select().from('products').where({id: id});
+      if(product.length === 0) return res.sendStatus(404);
+
       await db.delete().from('products').where({id : id});
       res.sendStatus(200);
+
     } catch (error) {
       error.name === 'ValidationError' ? res.sendStatus(400) : res.sendStatus(500);
       console.log(error);
